Clean up the order
This prepares the document for the next round of stuff we must handle.
The goal is to provide the text for transports next which explains how
you link stuff into the system. This is then followed by a section on
Mutations, which logically lead into the sections of Relay Modern
integration.