Fix loading of starships and vehicles
The bug is the following: Starships and Vehicles contain a transport
as a common table. To make stuff work out, you have to load these
correctly. We used 'Transport' as an atom rather than 'transport' and
then mnesia tells us the table doesn't exist.

Second, we had a problem when trying to type resolve over an abstract
Node to a Starship (and Vehicle). Add the correct type resolver code
for this.